Completed Readme.md file
Changed title and added a description details what the project is about and how it's used. Added a picture for functionality. Then, added a detailed usage section regarding the project, with an embedded link to the youtube video showcasing it's full functionality. Lastly, added credits section to jest and inquirer